Extension: add commentator array, schedule.js
